requirements.txt
Django==5.0.4       
djangorestframework==3.16.1
django-cors-headers==4.9.0

python-dotenv==1.2.1

openai==2.23.0
transformers==5.2.0
sentence-transformers==5.2.3
scikit-learn==1.8.0